requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coreapi
coreschema
cryptography
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.19
django-cors-headers
django-crispy-forms
django-import-export
django-ranged-response==0.2.0
django-templated-mail==1.1.1
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
et-xmlfile
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl
pathspec
Pillow
platformdirs==2.4.1
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.0
tomli==2.0.1
typing-extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
